requirements.txt
slack_sdk
google-generativeai
duckduckgo_search
python-dotenv
streamlit
pyyaml
PyGithub
